Update lambda function to fetch and build arbitrary revisions
Note that this isn't sound with respect to dependencies: in particular, it will
be building zephyr and its dependencies from the container, using the build
toolchain from the container. To do better than this, we could look at building
using nix, but at that point we probably no longer want to be using the lambda
approach.